+++ changelog.md
@@ -8,6 +8,7 @@ Changelog
 - #32 improved resource handling
 - removed DEV_MODE again
 - #17 working on discovery and decided to stick with simple config file
+- #22 added image overlay
 
 0.3.0
 -----